+++ lib/dunolint_engine/src/file_kind.ml
@@ -19,52 +19,23 @@
 (*  <http://www.gnu.org/licenses/> and <https://spdx.org>, respectively.         *)
 (*********************************************************************************)
 
-module Handler = Handler
-module Linter = Linter
-module Linters = Linters
-module Sexp_handler = Sexp_handler
-module Stanza_linter = Stanza_linter
+type t = Unix.file_kind =
+  | S_REG
+  | S_DIR
+  | S_CHR
+  | S_BLK
+  | S_LNK
+  | S_FIFO
+  | S_SOCK
+[@@deriving enumerate]
 
-module Stanza = struct
-  type 'a t =
-    { stanza : 'a
-    ; path : Relative_path.t
-    ; original_sexp : Sexp.t
-    ; sexps_rewriter : Sexps_rewriter.t
-    ; linter : Linter.t
-    }
-end
-
-module type S = Dunolinter_intf.S with type 'a stanza := 'a Stanza.t
-
-let match_stanza (t : _ Stanza.t) = t.stanza
-let path (t : _ Stanza.t) = t.path
-let original_sexp (t : _ Stanza.t) = t.original_sexp
-let sexps_rewriter (t : _ Stanza.t) = t.sexps_rewriter
-let linter (t : _ Stanza.t) = t.linter
-
-let eval_path ~path ~predicate =
-  match (predicate : Dunolint.Predicate.t) with
-  | `dune _ | `dune_project _ -> Dunolint.Trilang.Undefined
-  | `path condition ->
-    Blang.eval condition (function
-      | `equals value -> Relative_path.equal path value
-      | `glob glob -> Dunolint.Glob.test glob (Relative_path.to_string path))
-    |> Dunolint.Trilang.const
+let to_string t =
+  match t with
+  | S_REG -> "Regular file"
+  | S_DIR -> "Directory"
+  | S_CHR -> "Character device"
+  | S_BLK -> "Block device"
+  | S_LNK -> "Symbolic link"
+  | S_FIFO -> "Named pipe"
+  | S_SOCK -> "Socket"
 ;;
-
-module Private = struct
-  module Stanza = struct
-    module For_create = struct
-      type nonrec 'a t = 'a Stanza.t =
-        { stanza : 'a
-        ; path : Relative_path.t
-        ; original_sexp : Sexp.t
-        ; sexps_rewriter : Sexps_rewriter.t
-        ; linter : Linter.t
-        }
-    end
-
-    let create t = t
-  end
-end