+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ### Added
 
+- Added command `tools lint-file` to ease editor integration (#39, #40, #46, @mbarbin, requested by and with help from @arvidj).
 - Added support for `library.modes` (#23, #27, @mbarbin, requested by @raphael-proust)
 - Document release process and current state (#25, @mbarbin).
 - Added coverage and regression tests (#18, #19, #30, #33, #34, @mbarbin).